Improve the e2e tests
